+++ internal/repos/postgresql/work_item_project_demo.go
@@ -22,7 +22,7 @@ func NewDemoWorkItem() *DemoWorkItem {
 
 var _ repos.DemoWorkItem = (*DemoWorkItem)(nil)
 
-func (u *DemoWorkItem) ByID(ctx context.Context, d db.DBTX, id int64, opts ...db.WorkItemSelectConfigOption) (*db.WorkItem, error) {
+func (u *DemoWorkItem) ByID(ctx context.Context, d db.DBTX, id int, opts ...db.WorkItemSelectConfigOption) (*db.WorkItem, error) {
 	extraOpts := []db.WorkItemSelectConfigOption{db.WithWorkItemJoin(db.WorkItemJoins{DemoWorkItem: true})}
 	return db.WorkItemByWorkItemID(ctx, d, id, (append(extraOpts, opts...))...)
 }
@@ -44,7 +44,7 @@ func (u *DemoWorkItem) Create(ctx context.Context, d db.DBTX, params repos.DemoW
 	return workItem, nil
 }
 
-func (u *DemoWorkItem) Update(ctx context.Context, d db.DBTX, id int64, params repos.DemoWorkItemUpdateParams) (*db.WorkItem, error) {
+func (u *DemoWorkItem) Update(ctx context.Context, d db.DBTX, id int, params repos.DemoWorkItemUpdateParams) (*db.WorkItem, error) {
 	workItem, err := u.ByID(ctx, d, id)
 	if err != nil {
 		return nil, fmt.Errorf("could not get workItem by id: %w", parseErrorDetail(err))
@@ -73,7 +73,7 @@ func (u *DemoWorkItem) Update(ctx context.Context, d db.DBTX, id int64, params r
 	return workItem, err
 }
 
-func (u *DemoWorkItem) Delete(ctx context.Context, d db.DBTX, id int64) (*db.WorkItem, error) {
+func (u *DemoWorkItem) Delete(ctx context.Context, d db.DBTX, id int) (*db.WorkItem, error) {
 	workItem, err := u.ByID(ctx, d, id)
 	if err != nil {
 		return nil, fmt.Errorf("could not get workItem: %w", parseErrorDetail(err))
@@ -87,7 +87,7 @@ func (u *DemoWorkItem) Delete(ctx context.Context, d db.DBTX, id int64) (*db.Wor
 	return workItem, err
 }
 
-func (u *DemoWorkItem) Restore(ctx context.Context, d db.DBTX, id int64) (*db.WorkItem, error) {
+func (u *DemoWorkItem) Restore(ctx context.Context, d db.DBTX, id int) (*db.WorkItem, error) {
 	var err error
 	workItem := &db.WorkItem{
 		WorkItemID: id,
