+++ internal/repos/reposwrappers/workitem_with_retry.gen.go
@@ -29,7 +29,7 @@ func NewWorkItemWithRetry(base repos.WorkItem, retryCount int, retryInterval tim
 }
 
 // ByID implements repos.WorkItem
-func (_d WorkItemWithRetry) ByID(ctx context.Context, d db.DBTX, id int64, opts ...db.WorkItemSelectConfigOption) (wp1 *db.WorkItem, err error) {
+func (_d WorkItemWithRetry) ByID(ctx context.Context, d db.DBTX, id int, opts ...db.WorkItemSelectConfigOption) (wp1 *db.WorkItem, err error) {
 	wp1, err = _d.WorkItem.ByID(ctx, d, id, opts...)
 	if err == nil || _d._retryCount < 1 {
 		return
